requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.5
django-bootstrap4==22.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-nested-admin==3.4.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
dynaconf==3.1.9
gunicorn==20.1.0
hvac==0.11.2
idna==3.3
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
knox==0.1.14
loguru==0.6.0
MarkupSafe==2.1.1
Pillow==9.1.1
psycopg2==2.9.3
pyaml==21.10.1
pycparser==2.21
python-decouple==3.6
python-monkey-business==1.0.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
validators==0.20.0
whitenoise==6.2.0
